feat(blog): refactor BlogPage to use BlogPostsSection for improved structure and add category filtering functionality
